Don't try to save or remove a null session.
Null pointer exception reported by http://github.com/scottbarr occurs
in Tomcat 6. While I don't have that installed, line 890 of http://is.gd/Ni2
makes it pretty clear.  This fix first determines if a session exists,
and if it doesn't then we simply exit.

Also fix a typo in one of the comments.

Signed-off-by: Tim Shadel <github@timshadel.com>